+++ caffe2/CMakeLists.txt
@@ -13,6 +13,7 @@ add_subdirectory(cuda_rtc)
 add_subdirectory(db)
 add_subdirectory(distributed)
 # add_subdirectory(experiments) # note, we may remove this folder at some point
+add_subdirectory(ideep)
 add_subdirectory(image)
 add_subdirectory(video)
 add_subdirectory(mkl)
